requirements.txt
# python==3.9.20
# torch==2.0.0+cu117
# torchaudio==2.0.0
# torchvision==0.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
axial_positional_embedding==0.2.1
backcall==0.2.0
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
CoLT5-attention==0.11.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
einops==0.8.0
entrypoints==0.4
executing==2.1.0
filelock==3.13.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gluonts==0.15.1
gmpy2==2.1.2
huggingface-hub==0.25.2
idna==3.7
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter-client==7.3.4
jupyter_core==5.7.2
kiwisolver==1.4.7
local-attention==1.9.15
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.2.1
numpy==1.24.4
packaging==24.1
pandas==1.5.3
parso==0.8.4
patool==1.15.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
product_key_memory==0.2.11
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
reformer-pytorch==1.4.4
requests==2.32.3
safetensors==0.4.5
scikit-base==0.8.3
scikit-learn==1.5.2
scipy==1.11.3
setuptools==75.1.0
six==1.16.0
sktime==0.33.1
stack-data==0.6.2
sympy==1.13.2
threadpoolctl==3.5.0
timm==1.0.9
toolz==0.12.1
tornado==6.1
tqdm==4.66.5
traitlets==5.14.3
triton==2.0.0
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.3
visionts==0.2.0
wcwidth==0.2.13
wheel==0.44.0
xxhash==3.5.0
yarl==1.15.0
zipp==3.20.2
